+++ Genesis/Changes.md
@@ -5,6 +5,8 @@
 new mutable mouse event handlers.
 - Added absoluteMousePosition to all MouseEvents
 - Path no longer extends DistanceLike
+- Renamed RGB, RGBLike and RGBChannel to Rgb RgbLike and RgbChannel
+- Renamed HSL and HSLLike to Hsl and HslLike
 ### New Features
 - Added global mouse and keyboard event handling with GlobalMouseEventHandler and GlobalKeyboardEventHandler objects
 - Added SegmentedPath for path / animation based on a set of values